refactor: mvc 형태로 전환
