+++ src/main/java/stringcalcaulator/StringCalculatorMain.java
@@ -2,17 +2,13 @@ package stringcalcaulator;
 
 import java.util.Scanner;
 
+import stringcalcaulator.controller.StringCalcController;
+import stringcalcaulator.domain.StringCalculator;
+
 public class StringCalculatorMain {
 
     public static void main(String[] args) {
-        Scanner scanner = new Scanner(System.in);
-
-        System.out.println("계산할 문자열을 입력하세요.");
-
-        StringCalculator calc = new StringCalculator();
-        int result = calc.calculate(scanner.nextLine());
-
-        System.out.println(result);
+        new StringCalcController().run();
     }
 
 }
