+++ src/main/java/stringcalcaulator/domain/token/ExpressionToken.java
@@ -1,0 +1,7 @@
+package stringcalcaulator.domain.token;
+
+import stringcalcaulator.domain.ExpressionTokenVisitor;
+
+public interface ExpressionToken {
+    void accept(ExpressionTokenVisitor visitor);
+}
