+++ src/main/java/stringcalcaulator/domain/token/ExpressionTokenResolver.java
@@ -1,4 +1,4 @@
-package stringcalcaulator.token;
+package stringcalcaulator.domain.token;
 
 public interface ExpressionTokenResolver {
     boolean supports(String token);
