+++ src/main/java/stringcalcaulator/domain/token/ExpressionTokenResolvers.java
@@ -1,4 +1,4 @@
-package stringcalcaulator.token;
+package stringcalcaulator.domain.token;
 
 import java.util.Arrays;
 import java.util.List;
