--- src/main/java/stringcalcaulator/token/ExpressionToken.java
@@ -1,7 +1,0 @@
-package stringcalcaulator.token;
-
-import stringcalcaulator.ExpressionTokenVisitor;
-
-public interface ExpressionToken {
-    void accept(ExpressionTokenVisitor visitor);
-}
