+++ src/main/java/stringcalcaulator/view/StringCalculatorResultView.java
@@ -1,0 +1,9 @@
+package stringcalcaulator.view;
+
+public class StringCalculatorResultView {
+
+    public void printResult(int result) {
+        System.out.println(result);
+    }
+
+}
