+++ src/test/java/stringcalcaulator/domain/StringCalculatorTest.java
@@ -1,7 +1,5 @@
-package stringcalcaulator;
+package stringcalcaulator.domain;
 
-import org.assertj.core.api.Assertions;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
